This is all a little ugly and semi-working and maybe needs a
This is all a little ugly and semi-working and maybe needs a
rethink/rewrite. But it shouldn't break anything that works now, so here it
is. It mostly works if you don't test any boundary cases.

Sun Jul 12 22:57:55 1998  Havoc Pennington  <hp@pobox.com>

* gnome-app-util.c: Implementations for interactive appbar
and appbar progress bar.
* gnome-app-util.h: Change set_progress to use gdouble rather
than gfloat.
* gnome-appbar.h, gnome-appbar.c: Try implementing interactive
minibuffer with a GtkEntry hack. It doesn't work so good; think of
it as a prototype.
* gnome-types.h: GnomePreferencesType enum
* gnome-dialog-util.c: Use set_modal instead of grab_add for modal
dialogs; this fixes a bug (the "modal" dialogs weren't).


Sun Jul 12 23:01:56 1998  Havoc Pennington  <hp@pobox.com>

* testgnome.c (create_app_util): Added a test for the
GnomeAppBar, gnome-app-util, and gnome-dialog-util (they're
all interrelated, so it's just one test window). This is also
a decent demo for what these functions do (or don't).

Also added a "Reload preferences" button to load preferences
and observe the effect on the gnome libraries.

+++ libgnomeui/gnome-app-util.h
@@ -75,7 +75,7 @@ gnome_app_request_password (GnomeApp * app, const gchar * prompt,
 /* ========================================================== */
 
 /* Returns the fraction of the task done at a given time. */
-typedef gfloat (* GnomeAppProgressFunc) (gpointer data);
+typedef gdouble (* GnomeAppProgressFunc) (gpointer data);
 
 /* What to call if the operation is canceled. */
 typedef void (* GnomeAppProgressCancelFunc) (gpointer data);
@@ -106,7 +106,7 @@ gnome_app_progress_manual (GnomeApp * app,
 
 /* Only makes sense with manual. */
 void 
-gnome_app_set_progress (GnomeAppProgressKey key, gfloat percent);
+gnome_app_set_progress (GnomeAppProgressKey key, gdouble percent);
 
 /* Call this when the progress meter should go away. Automatically 
    called if progress is cancelled. */
